# Scheduling Profiler: Move preprocessing to web worker and add loading indicator