--- a/packages/react-devtools-scheduling-profiler/src/App.js
+++ b/packages/react-devtools-scheduling-profiler/src/App.js
@@ -14,7 +14,6 @@ import '@reach/tooltip/styles.css';
 
 import * as React from 'react';
 
-import {ModalDialogContextController} from 'react-devtools-shared/src/devtools/views/ModalDialog';
 import {SchedulingProfiler} from './SchedulingProfiler';
 import {useBrowserTheme} from './hooks';
 
@@ -25,12 +24,10 @@ export default function App() {
   useBrowserTheme();
 
   return (
-    <ModalDialogContextController>
-      <div className={styles.DevTools}>
-        <div className={styles.TabContent}>
-          <SchedulingProfiler />
-        </div>
+    <div className={styles.DevTools}>
+      <div className={styles.TabContent}>
+        <SchedulingProfiler />
       </div>
-    </ModalDialogContextController>
+    </div>
   );
 }