--- a/packages/react-devtools-scheduling-profiler/src/App.js
+++ b/packages/react-devtools-scheduling-profiler/src/App.js
@@ -15,13 +15,14 @@ import '@reach/tooltip/styles.css';
 import * as React from 'react';
 
 import {SchedulingProfiler} from './SchedulingProfiler';
-import {useBrowserTheme} from './hooks';
+import {useBrowserTheme, useDisplayDensity} from './hooks';
 
 import styles from './App.css';
 import 'react-devtools-shared/src/devtools/views/root.css';
 
 export default function App() {
   useBrowserTheme();
+  useDisplayDensity();
 
   return (
     <div className={styles.DevTools}>