--- a/packages/react-devtools-scheduling-profiler/src/CanvasPage.js
+++ b/packages/react-devtools-scheduling-profiler/src/CanvasPage.js
@@ -16,6 +16,7 @@ import type {
   ReactProfilerData,
   ReactMeasure,
 } from './types';
+import type {ProfilerDataResource} from './SchedulingProfiler';
 
 import * as React from 'react';
 import {
@@ -62,10 +63,11 @@ import styles from './CanvasPage.css';
 const CONTEXT_MENU_ID = 'canvas';
 
 type Props = {|
-  profilerData: ReactProfilerData,
+  dataResource: ProfilerDataResource,
 |};
 
-function CanvasPage({profilerData}: Props) {
+function CanvasPage({dataResource}: Props) {
+  const profilerData = dataResource.read();
   return (
     <div
       className={styles.CanvasPage}