--- a/packages/react-devtools-scheduling-profiler/src/ImportButton.js
+++ b/packages/react-devtools-scheduling-profiler/src/ImportButton.js
@@ -7,8 +7,8 @@
  * @flow
  */
 
-import type {TimelineEvent} from '@elg/speedscope';
 import type {ReactProfilerData} from './types';
+import type {ImportWorkerOutputData} from './import-worker/import.worker';
 
 import * as React from 'react';
 import {useCallback, useContext, useRef} from 'react';
@@ -17,10 +17,8 @@ import Button from 'react-devtools-shared/src/devtools/views/Button';
 import ButtonIcon from 'react-devtools-shared/src/devtools/views/ButtonIcon';
 import {ModalDialogContext} from 'react-devtools-shared/src/devtools/views/ModalDialog';
 
-import preprocessData from './utils/preprocessData';
-import {readInputData} from './utils/readInputData';
-
 import styles from './ImportButton.css';
+import ImportWorker from './import-worker/import.worker';
 
 type Props = {|
   onDataImported: (profilerData: ReactProfilerData) => void,
@@ -30,33 +28,39 @@ export default function ImportButton({onDataImported}: Props) {
   const inputRef = useRef<HTMLInputElement | null>(null);
   const {dispatch: modalDialogDispatch} = useContext(ModalDialogContext);
 
-  const handleFiles = useCallback(async () => {
+  const handleFiles = useCallback(() => {
     const input = inputRef.current;
     if (input === null) {
       return;
     }
 
     if (input.files.length > 0) {
-      try {
-        const readFile = await readInputData(input.files[0]);
-        const events: TimelineEvent[] = JSON.parse(readFile);
-        if (events.length > 0) {
-          onDataImported(preprocessData(events));
+      const worker: Worker = new (ImportWorker: any)();
+      worker.onmessage = function(event) {
+        const data = ((event.data: any): ImportWorkerOutputData);
+        switch (data.status) {
+          case 'SUCCESS':
+            onDataImported(data.processedData);
+            break;
+          case 'ERROR':
+            modalDialogDispatch({
+              type: 'SHOW',
+              title: 'Import failed',
+              content: (
+                <>
+                  <div>The profiling data you selected cannot be imported.</div>
+                  {data.error !== null && (
+                    <div className={styles.ErrorMessage}>
+                      {data.error.message}
+                    </div>
+                  )}
+                </>
+              ),
+            });
+            break;
         }
-      } catch (error) {
-        modalDialogDispatch({
-          type: 'SHOW',
-          title: 'Import failed',
-          content: (
-            <>
-              <div>The profiling data you selected cannot be imported.</div>
-              {error !== null && (
-                <div className={styles.ErrorMessage}>{error.message}</div>
-              )}
-            </>
-          ),
-        });
-      }
+      };
+      worker.postMessage({file: input.files[0]});
     }
 
     // Reset input element to allow the same file to be re-imported