--- a/packages/react-devtools-scheduling-profiler/src/SchedulingProfiler.js
+++ b/packages/react-devtools-scheduling-profiler/src/SchedulingProfiler.js
@@ -42,6 +42,7 @@ function createDataResourceFromImportedFile(file: File): ProfilerDataResource {
               reject(data.error);
               break;
           }
+          worker.terminate();
         };
         worker.postMessage({file});
       });