--- a/packages/react-devtools-scheduling-profiler/src/SchedulingProfiler.js
+++ b/packages/react-devtools-scheduling-profiler/src/SchedulingProfiler.js
@@ -13,37 +13,40 @@ import type {ImportWorkerOutputData} from './import-worker/import.worker';
 
 import * as React from 'react';
 import {Suspense, useCallback, useState} from 'react';
-import {unstable_batchedUpdates as batchedUpdates} from 'react-dom';
 import {createResource} from 'react-devtools-shared/src/devtools/cache';
-
-import ImportButton from './ImportButton';
 import ReactLogo from 'react-devtools-shared/src/devtools/views/ReactLogo';
 
+import ImportButton from './ImportButton';
 import CanvasPage from './CanvasPage';
 import ImportWorker from './import-worker/import.worker';
 
 import profilerBrowser from './assets/profilerBrowser.png';
 import styles from './SchedulingProfiler.css';
 
-export type ProfilerDataResource = Resource<void, File, ReactProfilerData>;
+type DataResource = Resource<void, File, ReactProfilerData | Error>;
 
-function createDataResourceFromImportedFile(file: File): ProfilerDataResource {
+function createDataResourceFromImportedFile(file: File): DataResource {
   return createResource(
     () => {
-      return new Promise<ReactProfilerData>((resolve, reject) => {
+      return new Promise<ReactProfilerData | Error>((resolve, reject) => {
         const worker: Worker = new (ImportWorker: any)();
+
         worker.onmessage = function(event) {
           const data = ((event.data: any): ImportWorkerOutputData);
           switch (data.status) {
             case 'SUCCESS':
               resolve(data.processedData);
               break;
-            case 'ERROR':
+            case 'INVALID_PROFILE_ERROR':
+              resolve(data.error);
+              break;
+            case 'UNEXPECTED_ERROR':
               reject(data.error);
               break;
           }
           worker.terminate();
         };
+
         worker.postMessage({file});
       });
     },
@@ -53,23 +56,11 @@ function createDataResourceFromImportedFile(file: File): ProfilerDataResource {
 }
 
 export function SchedulingProfiler(_: {||}) {
-  const [dataResource, setDataResource] = useState<ProfilerDataResource | null>(
-    null,
-  );
+  const [dataResource, setDataResource] = useState<DataResource | null>(null);
 
-  // A key to safely reset the error boundary
-  // See https://github.com/bvaughn/react-error-boundary/issues/23#issuecomment-425470511
-  const [importErrorBoundaryKey, setImportErrorBoundaryKey] = useState(0);
-
-  const handleFileSelect = useCallback(
-    (file: File) => {
-      batchedUpdates(() => {
-        setDataResource(createDataResourceFromImportedFile(file));
-        setImportErrorBoundaryKey(importErrorBoundaryKey + 1);
-      });
-    },
-    [importErrorBoundaryKey],
-  );
+  const handleFileSelect = useCallback((file: File) => {
+    setDataResource(createDataResourceFromImportedFile(file));
+  }, []);
 
   return (
     <div className={styles.SchedulingProfiler}>
@@ -82,13 +73,12 @@ export function SchedulingProfiler(_: {||}) {
       </div>
       <div className={styles.Content}>
         {dataResource ? (
-          <ImportErrorBoundary
-            key={importErrorBoundaryKey}
-            onFileSelect={handleFileSelect}>
-            <Suspense fallback={<ProcessingData />}>
-              <CanvasPage dataResource={dataResource} />
-            </Suspense>
-          </ImportErrorBoundary>
+          <Suspense fallback={<ProcessingData />}>
+            <DataResourceComponent
+              dataResource={dataResource}
+              onFileSelect={handleFileSelect}
+            />
+          </Suspense>
         ) : (
           <Welcome onFileSelect={handleFileSelect} />
         )}
@@ -97,11 +87,7 @@ export function SchedulingProfiler(_: {||}) {
   );
 }
 
-type WelcomeProps = {|
-  onFileSelect: (file: File) => void,
-|};
-
-const Welcome = ({onFileSelect}: WelcomeProps) => (
+const Welcome = ({onFileSelect}: {|onFileSelect: (file: File) => void|}) => (
   <div className={styles.EmptyStateContainer}>
     <div className={styles.ScreenshotWrapper}>
       <img
@@ -126,33 +112,34 @@ const ProcessingData = () => (
   </div>
 );
 
-class ImportErrorBoundary extends React.Component {
-  state: {|error: Error | null|} = {error: null};
+const CouldNotLoadProfile = ({error, onFileSelect}) => (
+  <div className={styles.EmptyStateContainer}>
+    <div className={styles.Header}>Could not load profile</div>
+    {error.message && (
+      <div className={styles.Row}>
+        <div className={styles.ErrorMessage}>{error.message}</div>
+      </div>
+    )}
+    <div className={styles.Row}>
+      Try importing
+      <ImportButton onFileSelect={onFileSelect} />
+      another Chrome performance profile.
+    </div>
+  </div>
+);
 
-  static getDerivedStateFromError(error) {
-    return {error};
+const DataResourceComponent = ({
+  dataResource,
+  onFileSelect,
+}: {|
+  dataResource: DataResource,
+  onFileSelect: (file: File) => void,
+|}) => {
+  const dataOrError = dataResource.read();
+  if (dataOrError instanceof Error) {
+    return (
+      <CouldNotLoadProfile error={dataOrError} onFileSelect={onFileSelect} />
+    );
   }
-
-  render() {
-    const {children, onFileSelect} = this.props;
-    const {error} = this.state;
-
-    if (error) {
-      return (
-        <div className={styles.EmptyStateContainer}>
-          <div className={styles.Header}>Could not load profile</div>
-          <div className={styles.Row}>
-            <div className={styles.ErrorMessage}>{error.message}</div>
-          </div>
-          <div className={styles.Row}>
-            Try importing
-            <ImportButton onFileSelect={onFileSelect} />
-            another Chrome performance profile.
-          </div>
-        </div>
-      );
-    }
-
-    return children;
-  }
-}
+  return <CanvasPage profilerData={dataOrError} />;
+};