<!--
  Thanks for submitting a pull request!
  We appreciate you spending the time to work on these changes. Please provide enough information so that others can review your pull request. The three fields below are mandatory.

  Before submitting a pull request, please make sure the following is done:

  1. Fork [the repository](https://github.com/facebook/react) and create your branch from `master`.
  2. Run `yarn` in the repository root.
  3. If you've fixed a bug or added code that should be tested, add tests!
  4. Ensure the test suite passes (`yarn test`). Tip: `yarn test --watch TestName` is helpful in development.
  5. Run `yarn test-prod` to test in the production environment. It supports the same options as `yarn test`.
  6. If you need a debugger, run `yarn debug-test --watch TestName`, open `chrome://inspect`, and press "Inspect".
  7. Format your code with [prettier](https://github.com/prettier/prettier) (`yarn prettier`).
  8. Make sure your code lints (`yarn lint`). Tip: `yarn linc` to only check changed files.
  9. Run the [Flow](https://flowtype.org/) type checks (`yarn flow`).
  10. If you haven't already, complete the CLA.

  Learn more about contributing: https://reactjs.org/docs/how-to-contribute.html
-->


<!-- Explain the **motivation** for making this change. What existing problem does the pull request solve? -->

Builds on #19707 to further improve the UI/UX by:

1. Moving our expensive `preprocessData` call into a new web worker. This previously caused the UI to freeze for a significant amount of time (the facebook.com we have freezes my browser for >10 seconds).
1. Add a loading indicator. This `ProcessingData` component is copied from the React DevTools Profiler.

|Before: freezes, errors in modal |After: loading text + no freezing, full page errors|
|-|-|
|![old](https://user-images.githubusercontent.com/12784593/92119610-a1deed00-ee2a-11ea-91da-0fe3fab6dbe2.gif)|![general](https://user-images.githubusercontent.com/12784593/92116857-f718ff80-ee26-11ea-9662-2ba9dbc4d022.gif)|


cc @bvaughn, @kartikcho 

### Implementation notes

1. I've used Suspense to implement the loading indicator stuff. This is the first time I'm using Suspense to do anything real, so please let me know if I'm not doing anything right 😄
1. I've used the `Resource` implementation in `react-devtools-shared` without forking it.

### UI notes

It's possible to load a second profile while the first one is still loading. I think this is fine from both UI and technical perspectives: users may want to select another profile if they selected the wrong profile/the profile is taking too long to load, and our implementation only displays the latest *selected* profile.

This is what it looks like:
![loadover](https://user-images.githubusercontent.com/12784593/92117497-d2715780-ee27-11ea-9691-15fa486e7316.gif)

Here's what's happening in the gif:

1. Load large facebook.com profile.
1. While it's loading, I loaded a smaller profile that finished loading and was displayed.
1. I also loaded another profile that caused an error.
1. I then waited a few seconds for the facebook.com profile to finish loading (verified afterwards, where the flamechart shows that the loading completed). Verified that the facebook.com profile's data was not displayed.
1. We then wait for Chrome DevTools to process the performance profile.
1. Chrome DevTools' Performance tool shows that multiple workers were created and processed their profiles in parallel.