@taneliang 2020-09-03
Looks like the error handling is now broken with fast refresh.

![fast refresh overlay](https://user-images.githubusercontent.com/12784593/92128888-71e91700-ee35-11ea-8205-6326b211f64e.gif)

I haven't managed to find the cause yet, but here are some guesses:

1. I may not be using error boundaries correctly. The error boundary *is* catching the thrown error though. Maybe this is just not a good use case for error boundaries?
1. The error overlay may just be too eager.

Any idea, @bvaughn? I can look into this further tomorrow too.