@bvaughn 2020-09-04
I think this change is probably a good one, because freezing the browser when uploading a larger profile is not a good user experience, *but* it makes the overall upload process *noticeably slower*.

For example, I tested with a 95MB profile from Facebook (the largest one I had handy) and the existing app took about 1.6 seconds:
![Screen Shot 2020-09-04 at 10 19 14 AM](https://user-images.githubusercontent.com/29597/92249450-1dc34e80-ee98-11ea-9731-5b518061f979.png)

I used the same profile on (a production build of) this branch and it took about 10 seconds, most of which was in the worker:
![Screen Shot 2020-09-04 at 10 20 22 AM](https://user-images.githubusercontent.com/29597/92249545-477c7580-ee98-11ea-970d-93577db7e1a1.png)

But a decent chunk was still on main:
![Screen Shot 2020-09-04 at 10 20 54 AM](https://user-images.githubusercontent.com/29597/92249593-58c58200-ee98-11ea-8711-98fb3140e587.png)

How big is the profile you're testing with? What do the overall load durations look like for you?