@bvaughn 2020-09-04
Something else occurs to me that I don't know the answer for: Will this change make merging the new profiler into DevTools harder?