@taneliang 2020-09-04
> the existing app took about 1.6 seconds ... this branch and it took about 10 seconds

Ouch!

I've just tried using the 57MB facebook.com profile that you sent us, and it seems like there's a difference between the project from the fellowship and what we currently have on `master`. I'm not seeing any noticeable difference between `master` and this branch though. Which existing app are you referring to?

On https://react-scheduling-profiler.vercel.app/ (currently still deployed from the MLH Fellowship repo), I'm getting ~5 seconds

![image](https://user-images.githubusercontent.com/12784593/92250636-48e08900-eefe-11ea-8e7b-bb38b4a84b70.png)

On a prod build of the `master` branch, I'm getting ~12 seconds. Maybe something (Speedscope probably?) got deopted?

![image](https://user-images.githubusercontent.com/12784593/92250907-a7a60280-eefe-11ea-91b5-5e5c28eb6ff8.png)

On a prod build of this branch, I'm getting the same ~11 seconds

![image](https://user-images.githubusercontent.com/12784593/92251144-fa7fba00-eefe-11ea-8e7f-8d02cdab483d.png)