@bvaughn 2020-09-04
Oh, interesting. I was just testing against [react-scheduling-profiler.vercel.app](https://react-scheduling-profiler.vercel.app/) rather than switching between master and rebuilding.