@taneliang 2020-09-04
Yeah, I'll look into this further next week.

> Will this change make merging the new profiler into DevTools harder? I think extensions support workers but I've never done it before.

I don't know either 😆 I did think about this, and I figured that since there's very little code needed to use web workers it shouldn't be a big issue when integrating this into DevTools. In the worst case, we can probably just remove the web worker entirely.